Fully fleshed out version that should be all correct, just needs to run and train.

+++ MLP_Lab1/InputNode.cs
@@ -19,7 +19,7 @@ namespace MLP_Lab1
             this.outputs = outputs;
         }
 
-        public float NodeOutput(int testNum)
+        public double NodeOutput(int testNum)
         {
             return (float)outputs[testNum];
         }
